refactor: seperate MainStackNavigation
