+++ apps/enfpy-native/src/navigation/main-stack/MainStackNavigation.tsx
@@ -1,0 +1,18 @@
+import React from 'react';
+import {NavigationContainer} from '@react-navigation/native';
+import {createNativeStackNavigator} from '@react-navigation/native-stack';
+import Main from '../../screens/main/Main';
+import Sub from '../../screens/sub/Sub';
+
+const Stack = createNativeStackNavigator();
+
+export const MainStackNavigation = () => {
+  return (
+    <NavigationContainer>
+      <Stack.Navigator>
+        <Stack.Screen name="main" component={Main} />
+        <Stack.Screen name="sub" component={Sub} />
+      </Stack.Navigator>
+    </NavigationContainer>
+  );
+};
